# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Plantar Futuro API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://plantarfuturo.com/ws"
        },
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://swagger.local"
        }
    ],
    "paths": {
...
        "/api/sst/visitor": {
            "get": {
                "tags": [
                    "[SST] Visitantes"
                ],
                "summary": "Mostrar lista de visitantes",
                "operationId": "App\\Http\\Controllers\\Api\\VisitorController::index",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los visitantes."
                    },
                    "204": {
                        "description": "No hay resultados que mostrar."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "[SST] Visitantes"
                ],
                "summary": "Registrar Visitante",
                "operationId": "App\\Http\\Controllers\\Api\\VisitorController::store",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del visitante",
                                        "type": "string",
                                        "format": "string",
                                        "example": "Miguel"
                                    },
                                    "document": {
                                        "description": "Documento del visitante",
                                        "type": "string",
                                        "format": "string",
                                        "example": "1030256147"
                                    },
                                    "entity": {
                                        "description": "Entidad del visitante",
                                        "type": "string",
                                        "format": "string",
                                        "example": "ANLA"
                                    },
                                    "position": {
                                        "description": "Cargo del visitante",
                                        "type": "string",
                                        "format": "string",
                                        "example": "Supervisor"
                                    },
                                    "state": {
                                        "description": "Estado (1: Activo, 0: Inactivo)",
                                        "type": "boolean",
                                        "example": "1"
                                    },
                                    "project_id": {
                                        "description": "Id del proyecto",
                                        "type": "integer",
                                        "format": "int32",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visitante registrada."
                    },
                    "400": {
                        "description": "Request mal mandado."
                    },
                    "401": {
                        "description": "Ingreso no autorizado."
                    },
                    "405": {
                        "description": "metodo HTTP no permitido."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/sst/visitor/{id}": {
            "get": {
                "tags": [
                    "[SST] Visitantes"
                ],
                "summary": "Ver visitantes por proyecto",
                "operationId": "App\\Http\\Controllers\\Api\\VisitorController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del proyecto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los visitantes."
                    },
                    "204": {
                        "description": "No hay resultados que mostrar."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "[SST] Visitantes"
                ],
                "summary": "Editar Visitante",
                "operationId": "App\\Http\\Controllers\\Api\\VisitorController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id (Primary Key) del visitante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del visitante",
                                        "type": "string",
                                        "format": "string",
                                        "example": "Miguel"
                                    },
                                    "document": {
                                        "description": "Documento del visitante",
                                        "type": "string",
                                        "format": "string",
                                        "example": "1030256147"
                                    },
                                    "entity": {
                                        "description": "Entidad del visitante",
                                        "type": "string",
                                        "format": "string",
                                        "example": "ANLA"
                                    },
                                    "position": {
                                        "description": "Cargo del visitante",
                                        "type": "string",
                                        "format": "string",
                                        "example": "Supervisor"
                                    },
                                    "state": {
                                        "description": "Estado (1: Activo, 0: Inactivo)",
                                        "type": "boolean",
                                        "example": "1"
                                    },
                                    "project_id": {
                                        "description": "Id del proyecto",
                                        "type": "integer",
                                        "format": "int32",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visitante registrada."
                    },
                    "400": {
                        "description": "Request mal mandado."
                    },
                    "401": {
                        "description": "Ingreso no autorizado."
                    },
                    "405": {
                        "description": "metodo HTTP no permitido."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    }
}